pyproject.toml
[project]
name = "vmestetrenimbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "httpx>=0.28.1",
    "pydantic-settings>=2.7.1",
]

[dependency-groups]
api = [
    "alembic>=1.17.2",
    "fastapi>=0.125.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.7.1",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.3.0",
    "sqladmin>=0.19.0",
    "itsdangerous>=2.2.0",
]

bot = [
    "pytelegrambotapi>=4.29.1",
]

all = [
    # Все зависимости из api
    "alembic>=1.17.2",
    "fastapi>=0.125.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.7.1",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.3.0",
    "sqladmin>=0.19.0",
    "itsdangerous>=2.2.0",
    # Все зависимости из bot
    "pytelegrambotapi>=4.29.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
package = true
# Синхронизировать все группы зависимостей по умолчанию
default-groups = ["all"]

[project.scripts]
init-env = "scripts.init_env:main"
up-docker = "scripts.up_docker:main"

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    "backend/alembic/",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
